supports assignment expressions, and handles ParserErrors gracefully.
also abstracts away the usage of currentIndex in Parser, especially for the instances of current++ in the Parser code, where before the Parser code called current++ when current was already at the last token in list, the EOF token, which lead to java out of bounds runtime exceptions in those cases where current was advanced past the end of list.

and forgot to exit Lox before interpreting source code if detect any syntax errors.