fix: was not reverting to the right env after finishing executing block.
see Interpreter.executeBlock(), local.parent is not necessarily the same
as the previous environment that should revert to at the end of the code
execution, like in the case of functions calls, where local.parent would
be (would always be) globals, which is not necessarily the previous env
that should revert to after finishing executing the function's code.

e.g.
fun sayHi(name) {
  print "Hi" + name + "!";
}
fun sayHo(name) {
  sayHi(name); // after executing sayHi(), we should return to
               // sayHo()'s env, the env that was created when sayHo()'s
               // code started executing -- not the globals environment!
  print "Ho" + name + "!";
}