implements closures, supports functions as first-class citizens, with a major TODO to do.
TODO: maintain static scoping for all var and fun declarations, so that a var.eval or fun.exec
can only refer to var or fun declarations preceding those var or fun resolutions respectively.

closures are implemented by storing a reference to the function's immediately enclosing
environment, so that if the function definition relies on variables defined in that immediately enclosing
environment (extending to the chain of parents of that environment), function.call will have a reference
to that environment, function.call will have a reference to the function's closure.